TWMessageBarManager: Fix a bug where messages which were presented from the callback of a message wouldn't show up
This happened due to the fact that selecting a message will set the "messageVisible" boolean property to NO before calling the callback.
When a new message get's presented, this message will be instantly shown, due to the "messageVisible" property being NO. If no further message
is queued in the messageWindow will be set to nil and the callback message never shows up.

This can be fixed by setting "messageVisible" to NO after calling the callback, to avoid that the message will be shown instantly and instead remains in the queue.